+++ migration/idempotent/007-tracing-tags.sql
@@ -123,21 +123,21 @@ the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map
 
 DO $do$
 BEGIN
-	CREATE OPERATOR ps_trace.= (
+	CREATE OPERATOR ps_trace.%= (
 	    FUNCTION       = ps_trace.tag_v_eq,
 	    LEFTARG        = _ps_trace.tag_v,
 	    RIGHTARG       = _ps_trace.tag_v,
-        NEGATOR        = OPERATOR(ps_trace.<>),
+        NEGATOR        = OPERATOR(ps_trace.%<>),
 	    RESTRICT       = eqsel,
 	    JOIN           = eqjoinsel,
 	    HASHES, MERGES
 	);
 EXCEPTION
     WHEN SQLSTATE '42723' THEN -- operator already exists
-        EXECUTE format($q$ALTER OPERATOR ps_trace.=(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR ps_trace.%=(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
 END;
 $do$;
-COMMENT ON OPERATOR ps_trace.= (_ps_trace.tag_v, _ps_trace.tag_v)
+COMMENT ON OPERATOR ps_trace.%= (_ps_trace.tag_v, _ps_trace.tag_v)
 IS 'This operator is a part of custom ps_trace.tag_map type which is a wrapper for 
 the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map_* functions.';
 
@@ -209,20 +209,20 @@ the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map
 
 DO $do$
 BEGIN
-	CREATE OPERATOR ps_trace.<> (
+	CREATE OPERATOR ps_trace.%<> (
 	    FUNCTION       = ps_trace.tag_v_ne,
 	    LEFTARG        = _ps_trace.tag_v,
 	    RIGHTARG       = _ps_trace.tag_v,
-	    NEGATOR        = OPERATOR(ps_trace.=),
+	    NEGATOR        = OPERATOR(ps_trace.%=),
 	    RESTRICT       = neqsel,
 	    JOIN           = neqjoinsel
 	);
 EXCEPTION
     WHEN SQLSTATE '42723' THEN -- operator already exists
-        EXECUTE format($q$ALTER OPERATOR ps_trace.<>(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR ps_trace.%<>(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
 END;
 $do$;
-COMMENT ON OPERATOR ps_trace.<> (_ps_trace.tag_v, _ps_trace.tag_v)
+COMMENT ON OPERATOR ps_trace.%<> (_ps_trace.tag_v, _ps_trace.tag_v)
 IS 'This operator is a part of custom ps_trace.tag_map type which is a wrapper for 
 the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map_* functions.';
 
@@ -287,75 +287,75 @@ IS 'This function is a part of custom _ps_trace.tag_v type which is a wrapper fo
 
 DO $do$
 BEGIN
-	CREATE OPERATOR ps_trace.> (
+	CREATE OPERATOR ps_trace.%> (
 	    FUNCTION       = ps_trace.tag_v_gt,
 	    LEFTARG        = _ps_trace.tag_v,
 	    RIGHTARG       = _ps_trace.tag_v,
-	    NEGATOR        = OPERATOR(ps_trace.<=),
+	    NEGATOR        = OPERATOR(ps_trace.%<=),
 	    RESTRICT       = scalargtsel,
 	    JOIN           = scalargtjoinsel
 	);
 EXCEPTION
     WHEN SQLSTATE '42723' THEN -- operator already exists
-        EXECUTE format($q$ALTER OPERATOR ps_trace.>(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR ps_trace.%>(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
 END;
 $do$;
 
 
 DO $do$
 BEGIN
-	CREATE OPERATOR ps_trace.>= (
+	CREATE OPERATOR ps_trace.%>= (
 	    FUNCTION       = ps_trace.tag_v_ge,
 	    LEFTARG        = _ps_trace.tag_v,
 	    RIGHTARG       = _ps_trace.tag_v,
-	    NEGATOR        = OPERATOR(ps_trace.<),
+	    NEGATOR        = OPERATOR(ps_trace.%<),
 	    RESTRICT       = scalargesel,
 	    JOIN           = scalargejoinsel
 	);
 EXCEPTION
     WHEN SQLSTATE '42723' THEN -- operator already exists
-        EXECUTE format($q$ALTER OPERATOR ps_trace.>=(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR ps_trace.%>=(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
 END;
 $do$;
-COMMENT ON OPERATOR ps_trace.>= (_ps_trace.tag_v, _ps_trace.tag_v)
+COMMENT ON OPERATOR ps_trace.%>= (_ps_trace.tag_v, _ps_trace.tag_v)
 IS 'This operator is a part of custom ps_trace.tag_map type which is a wrapper for 
 the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map_* functions.';
 
 DO $do$
 BEGIN
-	CREATE OPERATOR ps_trace.< (
+	CREATE OPERATOR ps_trace.%< (
 	    FUNCTION       = ps_trace.tag_v_lt,
 	    LEFTARG        = _ps_trace.tag_v,
 	    RIGHTARG       = _ps_trace.tag_v,
-	    NEGATOR        = OPERATOR(ps_trace.>=),
+	    NEGATOR        = OPERATOR(ps_trace.%>=),
 	    RESTRICT       = scalarltsel,
 	    JOIN           = scalarltjoinsel
 	);
 EXCEPTION
     WHEN SQLSTATE '42723' THEN -- operator already exists
-        EXECUTE format($q$ALTER OPERATOR ps_trace.<(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR ps_trace.%<(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
 END;
 $do$;
-COMMENT ON OPERATOR ps_trace.< (_ps_trace.tag_v, _ps_trace.tag_v)
+COMMENT ON OPERATOR ps_trace.%< (_ps_trace.tag_v, _ps_trace.tag_v)
 IS 'This operator is a part of custom ps_trace.tag_map type which is a wrapper for 
 the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map_* functions.';
 
 DO $do$
 BEGIN
-	CREATE OPERATOR ps_trace.<= (
+	CREATE OPERATOR ps_trace.%<= (
 	    FUNCTION       = ps_trace.tag_v_le,
 	    LEFTARG        = _ps_trace.tag_v,
 	    RIGHTARG       = _ps_trace.tag_v,
-	    NEGATOR        = OPERATOR(ps_trace.>),
+	    NEGATOR        = OPERATOR(ps_trace.%>),
 	    RESTRICT       = scalarlesel,
 	    JOIN           = scalarlejoinsel
 	);
 EXCEPTION
     WHEN SQLSTATE '42723' THEN -- operator already exists
-        EXECUTE format($q$ALTER OPERATOR ps_trace.<=(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR ps_trace.%<=(_ps_trace.tag_v, _ps_trace.tag_v) OWNER TO %I$q$, current_user);
 END;
 $do$;
-COMMENT ON OPERATOR ps_trace.<= (_ps_trace.tag_v, _ps_trace.tag_v)
+COMMENT ON OPERATOR ps_trace.%<= (_ps_trace.tag_v, _ps_trace.tag_v)
 IS 'This operator is a part of custom ps_trace.tag_map type which is a wrapper for 
 the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map_* functions.';
 
@@ -364,17 +364,18 @@ the built-in jsonb. It is the same as its jsonb_ namesake, but relies on tag_map
 
 DO $do$
 BEGIN
-    CREATE OPERATOR CLASS btree_tag_v_ops
-    DEFAULT FOR TYPE _ps_trace.tag_v USING btree
-    AS
-            OPERATOR        1       ps_trace.<  ,
-            OPERATOR        2       ps_trace.<= ,
-            OPERATOR        3       ps_trace.=  ,
-            OPERATOR        4       ps_trace.>= ,
-            OPERATOR        5       ps_trace.>  ,
-            FUNCTION        1       _ps_trace.tag_v_cmp(_ps_trace.tag_v, _ps_trace.tag_v);
+    CREATE OPERATOR CLASS ps_trace.btree_tag_v_ops
+        DEFAULT FOR TYPE _ps_trace.tag_v USING btree
+        AS
+                OPERATOR        1       ps_trace.%<  ,
+                OPERATOR        2       ps_trace.%<= ,
+                OPERATOR        3       ps_trace.%=  ,
+                OPERATOR        4       ps_trace.%>= ,
+                OPERATOR        5       ps_trace.%>  ,
+                FUNCTION        1       _ps_trace.tag_v_cmp(_ps_trace.tag_v, _ps_trace.tag_v);
 EXCEPTION
     WHEN SQLSTATE '42710' THEN -- object already exists
-        EXECUTE format($q$ALTER OPERATOR CLASS public.btree_tag_v_ops USING btree OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR CLASS ps_trace.btree_tag_v_ops USING btree OWNER TO %I$q$, current_user);
+        EXECUTE format($q$ALTER OPERATOR FAMILY ps_trace.btree_tag_v_ops USING btree OWNER TO %I$q$, current_user);
 END;
 $do$;